+++ Sources/SwifterSwift/CoreGraphics/CGFloatExtensions.swift
@@ -9,12 +9,8 @@
 #if canImport(CoreGraphics)
 import CoreGraphics
 
-#if canImport(UIKit)
-import UIKit
-#endif
-
-#if canImport(Cocoa)
-import Cocoa
+#if canImport(Foundation)
+import Foundation
 #endif
 
 // MARK: - Properties
@@ -25,20 +21,24 @@ public extension CGFloat {
         return Swift.abs(self)
     }
 
+    #if canImport(Foundation)
     /// SwifterSwift: Ceil of CGFloat value.
     var ceil: CGFloat {
         return Foundation.ceil(self)
     }
+    #endif
 
     /// SwifterSwift: Radian value of degree input.
     var degreesToRadians: CGFloat {
         return .pi * self / 180.0
     }
 
+    #if canImport(Foundation)
     /// SwifterSwift: Floor of CGFloat value.
     var floor: CGFloat {
         return Foundation.floor(self)
     }
+    #endif
 
     /// SwifterSwift: Check if CGFloat is positive.
     var isPositive: Bool {