+++ Sources/SwifterSwift/SwiftStdlib/RangeReplaceableCollectionExtensions.swift
@@ -9,8 +9,7 @@
 // MARK: - Initializers
 public extension RangeReplaceableCollection {
 
-    /// Creates a new collection of a given size where for each position of the collection the value will be the result
-    /// of a call of the given expression.
+    /// SwifterSwift: Creates a new collection of a given size where for each position of the collection the value will be the result of a call of the given expression.
     ///
     ///     let values = Array(expression: "Value", count: 3)
     ///     print(values)
@@ -21,7 +20,7 @@ public extension RangeReplaceableCollection {
     ///   - count: The count of the collection.
     init(expression: @autoclosure () throws -> Element, count: Int) rethrows {
         self.init()
-        //swiftlint:disable:next empty_count
+        // swiftlint:disable:next empty_count
         if count > 0 {
             reserveCapacity(count)
             while self.count < count {
@@ -44,7 +43,7 @@ public extension RangeReplaceableCollection {
     /// - Parameter places: Number of places that the array be rotated. If the value is positive the end becomes the start, if it negative it's that start becom the end.
     /// - Returns: The new rotated collection.
     func rotated(by places: Int) -> Self {
-        //Inspired by: https://ruby-doc.org/core-2.2.0/Array.html#method-i-rotate
+        // Inspired by: https://ruby-doc.org/core-2.2.0/Array.html#method-i-rotate
         var copy = self
         return copy.rotate(by: places)
     }
@@ -131,4 +130,25 @@ public extension RangeReplaceableCollection {
         return Self(self[idx...])
     }
 
+    /// SwifterSwift: Remove all duplicate elements using KeyPath to compare.
+    ///
+    /// - Parameter path: Key path to compare, the value must be Equatable.
+    mutating func removeDuplicates<E: Equatable>(keyPath path: KeyPath<Element, E>) {
+        var items = [Element]()
+        removeAll { element -> Bool in
+            guard items.contains(where: { $0[keyPath: path] == element[keyPath: path] }) else {
+                items.append(element)
+                return false
+            }
+            return true
+        }
+    }
+
+    /// SwifterSwift: Remove all duplicate elements using KeyPath to compare.
+    ///
+    /// - Parameter path: Key path to compare, the value must be Hashable.
+    mutating func removeDuplicates<E: Hashable>(keyPath path: KeyPath<Element, E>) {
+        var set = Set<E>()
+        removeAll { !set.insert($0[keyPath: path]).inserted }
+    }
 }