+++ Tests/SharedTests/ColorExtensionsTests.swift
@@ -9,15 +9,7 @@
 import XCTest
 @testable import SwifterSwift
 
-#if canImport(UIKit) || canImport(Cocoa)
-
-#if os(macOS)
-import Cocoa
-public typealias Color = NSColor
-#else
-import UIKit
-public typealias Color = UIColor
-#endif
+#if canImport(AppKit) || canImport(UIKit)
 
 #if !os(watchOS)
 import CoreImage
@@ -77,7 +69,7 @@ final class ColorExtensionsTests: XCTestCase {
     }
 
     func testAlpha() {
-        var color: Color = Color.red
+        var color = Color.red
         XCTAssertEqual(color.alpha, 1.0)
 
         color = Color.white.withAlphaComponent(0.5)