+++ Tests/UIKitTests/UIGestureRecognizerExtensionsTests.swift
@@ -18,23 +18,23 @@ final class UIGestureRecognizerExtensionsTests: XCTestCase {
         let view = UIImageView()
         let tap = UITapGestureRecognizer()
 
-        //First Baseline Assertion
+        // First Baseline Assertion
         XCTAssert(view.gestureRecognizers == nil)
         XCTAssert(tap.view == nil)
 
         view.addGestureRecognizer(tap)
 
-        //Verify change
+        // Verify change
         XCTAssertFalse(view.gestureRecognizers == nil)
         XCTAssertFalse(tap.view == nil)
 
-        //Second Baseline Assertion
+        // Second Baseline Assertion
         XCTAssertFalse((view.gestureRecognizers?.count ?? 0) == 0)
         XCTAssertFalse(view.gestureRecognizers?.isEmpty ?? true)
 
         tap.removeFromView()
 
-        //Verify change
+        // Verify change
         XCTAssert((view.gestureRecognizers?.count ?? 1) == 0)
         XCTAssert(view.gestureRecognizers?.isEmpty ?? false)
         XCTAssert(tap.view == nil)